# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Logaroo",
        "description": "Documentação da API criada para o Desafio de backend developer da empresa Logaroo",
        "contact": {
            "name": "Wellington Santos",
            "email": "huske.xp@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registra um novo usuário",
                "description": "Esse endpoint registra um novo usuário",
                "operationId": "c7bb21a0d87049e161b8e4c9b0d909f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do Usuário",
                                        "type": "string",
                                        "example": "Exemplo Teste"
                                    },
                                    "email": {
                                        "description": "Email do Usuário",
                                        "type": "string",
                                        "example": "teste@example.org"
                                    },
                                    "password": {
                                        "description": "Senha do Usuário",
                                        "type": "string",
                                        "example": "#sdasdssdaAA@"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmação da senha do Usuário",
                                        "type": "string",
                                        "example": "#sdasdssdaAA@"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário Cadastrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário Cadastrado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos Incorretos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Esse email já foi cadastrado"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Recebe um token de autenticação",
                "description": "Esse endpoint retorna um token jwt de autenticação para o usuário",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email do Usuário",
                                        "type": "string",
                                        "example": "teste@example.org"
                                    },
                                    "password": {
                                        "description": "Senha do Usuário",
                                        "type": "string",
                                        "example": "#sdasdssdaAA@"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário logado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "2|MZEBxLy1zulPtND6brlf8GOPy57Q4DwYunlibXGj"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "string",
                                            "example": "3600"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciais  incorretas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "As credenciais enviadas estão incorretas"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Password incorreto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Recebe informações sobre o usuário autenticado",
                "description": "Esse endpoint retorna todas as informações sobre o usuário autenticado",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "Informações sobre o usuário autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Exemplo Teste"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "teste@example.org"
                                        },
                                        "email_verified_at": {
                                            "type": "string",
                                            "example": "2024-04-29T20:52:17.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-04-29T20:52:17.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-04-29T20:52:17.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciais  incorretas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "As credenciais enviadas estão incorretas"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Password incorreto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Remove todos os tokens do usuário",
                "description": "Esse endpoint faz o logout do usuário, revogando todos os tokens ativos dele",
                "operationId": "11f6351590f263ce3f8fd7598b962ade",
                "responses": {
                    "200": {
                        "description": "Tokens do usuário removidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todos os tokens do usuário removidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não Autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados enviados incorretos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados enviados estão incorretos"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}